check for relationships before deleting locality
git-svn-id: file:///repo/applications/FRED/trunk@10985 976ac4c7-77c1-46cc-9d05-16a3bd20634c

+++ nz/cri/gns/fred/util/FeatureUtil.java
@@ -333,6 +333,11 @@ public class FeatureUtil extends ModelUtil implements AuditedUtil {
 		if (feature.getFrNumber() != null)
 			feature.setFrNumber(null);
 		
+		if (!FREDUtil.isEmpty(feature.getRelationships())) {
+			Feature relFeature = ((Relationship)feature.getRelationships().iterator().next()).getFeature();
+			throw new IllegalStateException("Cannot delete this locality as is references in a relationship with " + FeatureUtil.getFeatureIdentifyingName(relFeature));
+		}
+		
 		Audit audit = feature.getAudit();
 
 		featureDAO.delete(feature);
